refactored to bigdecimal

+++ src/main/java/com/engine/dto/CandidateTradeDto.java
@@ -3,14 +3,16 @@ package com.engine.dto;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PositiveOrZero;
 
+import java.math.BigDecimal;
+
 public record CandidateTradeDto(
         @NotNull(message = "trade must be identifiable")
         String tradeName,
 
         @PositiveOrZero(message = "margin cant be negative")
-        int marginRequired,
+        BigDecimal marginRequired,
 
         @PositiveOrZero(message = "Pnl cant be negative")
-        int expectedPnl
+        BigDecimal expectedPnl
 ){}
