+++ src/main/java/com/engine/dto/RequestDto.java
@@ -6,12 +6,13 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.PositiveOrZero;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record RequestDto(
 
         @PositiveOrZero
-        int maxMargin,
+        BigDecimal maxMargin,
 
         @NotNull
         @Valid
