+++ src/main/java/com/engine/dto/ResponseDto.java
@@ -1,5 +1,6 @@
 package com.engine.dto;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 import java.util.UUID;
@@ -7,8 +8,8 @@ import java.util.UUID;
 public record ResponseDto(
         UUID requestId,
         List<CandidateTradeDto> selectedTrades,
-        int totalMarginRequired,
-        int totalExpectedPnl,
+        BigDecimal totalMarginRequired,
+        BigDecimal totalExpectedPnl,
         Instant createdAt
 ) {
 }
